re-organized file structure of league feature
