+++ lib/app/view/app_routing.dart
@@ -7,8 +7,8 @@ import 'package:ax_dapp/pages/farm/bloc/farm_bloc.dart';
 import 'package:ax_dapp/pages/farm/desktop_farm.dart';
 import 'package:ax_dapp/pages/farm/usecases/get_farm_data_use_case.dart';
 import 'package:ax_dapp/pages/landing_page/landing_page.dart';
-import 'package:ax_dapp/pages/league/bloc/league_bloc.dart';
-import 'package:ax_dapp/pages/league/desktop_leauge.dart';
+import 'package:ax_dapp/pages/league/league_search/bloc/league_bloc.dart';
+import 'package:ax_dapp/pages/league/league_search/views/desktop_leauge.dart';
 import 'package:ax_dapp/pages/league/repository/league_repository.dart';
 import 'package:ax_dapp/pages/trade/bloc/trade_page_bloc.dart';
 import 'package:ax_dapp/pages/trade/desktop_trade.dart';
