+++ lib/pages/league/league_draft/bloc/league_draft_bloc.dart
@@ -1,0 +1,13 @@
+import 'package:bloc/bloc.dart';
+import 'package:equatable/equatable.dart';
+
+part 'league_draft_event.dart';
+part 'league_draft_state.dart';
+
+class LeagueDraftBloc extends Bloc<LeagueDraftEvent, LeagueDraftState> {
+  LeagueDraftBloc() : super(LeagueDraftInitial()) {
+    on<LeagueDraftEvent>((event, emit) {
+      // TODO: implement event handler
+    });
+  }
+}
