+++ lib/pages/league/league_draft/bloc/league_draft_event.dart
@@ -1,0 +1,8 @@
+part of 'league_draft_bloc.dart';
+
+abstract class LeagueDraftEvent extends Equatable {
+  const LeagueDraftEvent();
+
+  @override
+  List<Object> get props => [];
+}
