+++ lib/pages/league/league_draft/bloc/league_draft_state.dart
@@ -1,0 +1,10 @@
+part of 'league_draft_bloc.dart';
+
+abstract class LeagueDraftState extends Equatable {
+  const LeagueDraftState();
+  
+  @override
+  List<Object> get props => [];
+}
+
+class LeagueDraftInitial extends LeagueDraftState {}
