+++ lib/pages/league/league_game/bloc/league_game_event.dart
@@ -1,0 +1,8 @@
+part of 'league_game_bloc.dart';
+
+abstract class LeagueGameEvent extends Equatable {
+  const LeagueGameEvent();
+
+  @override
+  List<Object> get props => [];
+}
