+++ lib/pages/league/league_game/bloc/league_game_state.dart
@@ -1,0 +1,10 @@
+part of 'league_game_bloc.dart';
+
+abstract class LeagueGameState extends Equatable {
+  const LeagueGameState();
+  
+  @override
+  List<Object> get props => [];
+}
+
+class LeagueGameInitial extends LeagueGameState {}
