+++ lib/pages/league/league_search/widgets/dialogs/rules_dialog.dart
@@ -1,4 +1,4 @@
-import 'package:ax_dapp/pages/league/bloc/league_bloc.dart';
+import 'package:ax_dapp/pages/league/league_search/bloc/league_bloc.dart';
 import 'package:ax_dapp/pages/league/models/league.dart';
 import 'package:ax_dapp/service/custom_styles.dart';
 import 'package:ax_dapp/wallet/bloc/wallet_bloc.dart';
