+++ lib/pages/league/league_search/widgets/league_card.dart
